pyproject.toml
[project]
name = "careeragent-ai"
version = "0.1.0"
description = "CareerAgent-AI: stateful agentic career automation with LangGraph + FastAPI + Gradio"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "pydantic>=2.0",
  "pydantic-settings>=2.2",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "langgraph>=0.2.0",
  "langchain-core>=0.2.0",
  "python-dotenv>=1.0",
  "loguru>=0.7.2",
  "tenacity>=8.2",
  "httpx>=0.27",
  "orjson>=3.10",
  "reportlab>=4.4.10",
  "jupyter>=1.1.1",
  "notebook>=7.5.3",
  "streamlit>=1.33",
  "twilio>=9.0",
  "pypdf>=6.7.1",
  "python-multipart>=0.0.22",
]

[project.optional-dependencies]
ui = ["gradio>=4.0"]
infra = [
  "azure-storage-blob>=12.19",
  "mlflow>=2.12",
]
ml = [
  "numpy>=1.26",
  "pandas>=2.1",
  "scikit-learn>=1.4",
  "rapidfuzz>=3.7",
]
docs = ["mkdocs-material>=9.5"]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/careeragent"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
  "mypy>=1.8",
  "types-requests>=2.31.0.20240106",
  "pre-commit>=3.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
